Add files via upload
This worksheet helps student calculate their CGPA with ease.
We are aware that for a student to have a recognized result, he/she must possess a Reg. No. Excel built-in function was use to execute this condition. i.e. once a student Reg. No is missing, definitely there won’t be a result for such student even if the name and grades were documented.
With respect to the above function, it is the job of the lecturer or student to ensure the Reg. No is correct for documentation as Excel calculates the GPA/CGPA once it recognizes a number in the Reg. No column.
Most importantly, to ensure accurate result of GPA/CGPA of a student after each successful session, it is mandatory that the order of input of Reg. No from the YEAR 1, remains the same till the final year.

This work sheet works perfectly for the Department of Statistics, FUTO, though any departments can still make use of it making slight adjustment in the course code and their units provided the university uses a 5.0 scale.